+++ packages/leancode_cubit_utils/test/utils/test_request_cubit.dart
@@ -4,17 +4,16 @@ import 'package:leancode_cubit_utils/leancode_cubit_utils.dart';
 import 'http_status_codes.dart';
 
 mixin RequestResultHandler<TOut>
-    on BaseRequestCubit<http.Response, String, TOut, int> {
+    on BaseRequestCubit<http.Response, String, int> {
   @override
-  Future<RequestState<TOut, int>> handleResult(http.Response result) async {
+  Future<RequestState<String, int>> handleResult(http.Response result) async {
     if (result.statusCode == StatusCode.ok.value) {
-      final data = map(result.body);
-      if (isEmpty(data)) {
+      if (result.body.isEmpty) {
         logger.warning('Query success but data is empty');
         return RequestEmptyState();
       }
       logger.info('Query success. Data: ${result.body}');
-      return RequestSuccessState(data);
+      return RequestSuccessState(result.body);
     } else {
       logger.severe('Query error. Status code: ${result.statusCode}');
       try {
@@ -29,7 +28,7 @@ mixin RequestResultHandler<TOut>
   }
 }
 
-class TestRequestCubit extends RequestCubit<http.Response, String, String, int>
+class TestRequestCubit extends RequestCubit<http.Response, String, int>
     with RequestResultHandler<String> {
   TestRequestCubit(
     super.loggerTag, {
@@ -40,19 +39,6 @@ class TestRequestCubit extends RequestCubit<http.Response, String, String, int>
 
   final http.Client client;
   final String id;
-
-  @override
-  String map(String data) {
-    if (data == 'Mapping fails') {
-      throw Exception('Mapping failed');
-    }
-    return 'Mapped $data';
-  }
-
-  @override
-  bool isEmpty(String data) {
-    return data.isEmpty;
-  }
 
   @override
   Future<http.Response> request() {
@@ -72,17 +58,11 @@ class TestRequestCubit extends RequestCubit<http.Response, String, String, int>
 }
 
 class TestArgsRequestCubit
-    extends ArgsRequestCubit<String, http.Response, String, String, int>
+    extends ArgsRequestCubit<String, http.Response, String, int>
     with RequestResultHandler<String> {
   TestArgsRequestCubit(super.loggerTag, {required this.client});
 
   final http.Client client;
-
-  @override
-  String map(String data) => data;
-
-  @override
-  bool isEmpty(String data) => data.isEmpty;
 
   @override
   Future<http.Response> request(String args) {